feat: Ensure stream position is always increased strictly monotonic...
for batch and bulk updates matching the global positions increasing index,
so that on subscriber side the events of one stream are delivered always in the
correct order to ensure consistency also on the query/read space.
Therefor if changes of a stream with multiple events are passed as single changes
instead of one bundled change with n events, it's merged to one change upfront
before executing the bacth or bulk (unnested) updated on the postgres eventstore.